requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2024.6.1
googletrans==3.0.0
gradio_client==1.2.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==1.0.5
httpx==0.26.0
huggingface-hub==0.24.5
hyperframe==5.2.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mistral_common==1.4.3
mwclient==0.11.0
mwparserfromhell==0.6.6
numpy==2.1.0
oauthlib==3.2.2
openai==1.54.5
packaging==24.1
pandas==2.2.2
pgpt_python==0.1.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rpds-py==0.20.0
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visitor==0.1.3
websockets==12.0
Werkzeug==3.0.3
WTForms==3.1.2
